[typescript] Start moving utility classes to Typescript.
