+++ to.etc.domui/src/main/resources/resources/js/odomui.js
@@ -159,6 +159,2661 @@ $(window).bind('beforeunload', function() {
 })(jQuery);
 
 /** WebUI helper namespace */
+var WebUI;
+if(WebUI === undefined)
+	WebUI = new Object();
+
+$.extend(WebUI, {
+	/**
+	 * can be set to true from server code with appendJavaScript so that the expired messages will not show and
+	 * block effortless refresh on class reload. Configurable in .developer.properties domui.hide-expired-alert.
+	 */
+	_hideExpiredMessage: false,
+
+	/**
+	 * Will be set by startPolling to define the poll interval.
+	 */
+	_pollInterval: 2500,
+
+	/**
+	 * When this is > 0, this keeps any page "alive" by sending an async
+	 */
+	_keepAliveInterval: 0,
+
+	_ignoreErrors: false,
+
+	setHideExpired: function() {
+		WebUI._hideExpiredMessage = true;
+	},
+
+
+	getInputFields : function(fields) {
+		// Collect all input, then create input.
+		var q1 = $("input").get();
+		for ( var i = q1.length; --i >= 0;) {
+			var t = q1[i];
+			if (t.type == 'file')
+				continue;
+			if (t.type == 'hidden' && !t.getAttribute('s')) // All hidden input nodes are created directly in browser java-script and because that are filtered out from server requests.
+				continue;
+
+			var val = undefined;
+			if (t.type == 'checkbox' || t.type == 'radio') {
+				val = t.checked ? "y" : "n";
+			} else {
+				val = t.value;
+			}
+
+			fields[t.id] = val;
+		}
+
+		var q1 = $("select").get();
+		for ( var i = q1.length; --i >= 0;) {
+			var sel = q1[i];
+			var val = undefined;
+			if (sel.selectedIndex != -1) {
+				val = sel.options[sel.selectedIndex].value;
+			}
+
+			if(val != undefined)
+				fields[sel.id] = val;
+		}
+		var q1 = $("textarea").get();
+		for ( var i = q1.length; --i >= 0;) {
+			var sel = q1[i];
+			if (sel.className == 'ui-ckeditor') {
+				//-- Locate the variable for this editor.
+				var editor = CKEDITOR.instances[sel.id];
+				if(null == editor)
+					throw "Cannot locate editor with id="+sel.id;
+				val = editor.getData();
+			} else {
+//				if($.browser.msie) { // The MS idiots remove newlines from value....
+//					val = sel.innerText;
+//					//alert("inner value="+sel.innerText);
+//				} else
+				val = sel.value;
+			}
+			fields[sel.id] = val;
+		}
+
+		//-- Handle all registered controls
+		var list = WebUI._inputFieldList;
+		for(var i = list.length; --i >= 0;) {
+			var item = list[i];
+			if(! document.getElementById(item.id)) {
+				//-- Node gone - remove input
+				list.splice(i, 1);
+			} else {
+				var data = item.control.getInputField();
+				fields[item.id] = data;
+			}
+		}
+
+		return fields;
+	},
+
+	_inputFieldList: [],
+
+	/**
+	 * This registers a non-html control as a source of input for {@link getInputFields}. The control
+	 * must have a method "getInputFields(fields: Map)" which defines the inputs to send for the control.
+	 */
+	registerInputControl: function(id, control) {
+		var list = WebUI._inputFieldList;
+		for(var i = list.length; --i >= 0;) {
+			var item = list[i];
+			if(item.id == id) {
+				item.control = control;
+				return;
+			}
+		}
+		list.push({id: id, control:control});
+	},
+
+	findInputControl: function(id) {
+		//-- return registered component by id, if not found returns null
+		var list = WebUI._inputFieldList;
+		for(var i = list.length; --i >= 0;) {
+			var item = list[i];
+			if(item.id == id && document.getElementById(item.id)) {
+				return item.control;
+			}
+		}
+		return null;
+	},
+
+	clicked : function(h, id, evt) {
+		//-- Trigger the before-clicked event on body
+		$(document.body).trigger("beforeclick", $("#"+id), evt);
+
+		// Collect all input, then create input.
+		var fields = new Object();
+		this.getInputFields(fields);
+		fields.webuia = "clicked";
+		fields.webuic = id;
+		fields["$pt"] = DomUIpageTag;
+		fields["$cid"] = DomUICID;
+		WebUI.cancelPolling();
+
+		//-- Do not call upward handlers too.
+		if(! evt)
+			evt = window.event;
+
+		// jal 20131107 Cancelling the event means that you cannot click items inside a clickable item
+		if(evt) {
+			evt.cancelBubble = true;
+			if(evt.stopPropagation)
+				evt.stopPropagation();
+		}
+		var e = $.event.fix(evt);		// Convert to jQuery event
+		//e.preventDefault(); // jal 20110216 DO NOT PREVENTDEFAULT- it will disable checkbox enable/disable
+
+		//-- add click-related parameters
+		fields._pageX = e.pageX;
+		fields._pageY = e.pageY;
+		fields._controlKey = e.ctrlKey == true;
+		fields._shiftKey = e.shiftKey == true;
+		fields._altKey = e.altKey == true;
+
+		$.ajax( {
+			url :DomUI.getPostURL(),
+			dataType :"*",
+			data :fields,
+			cache :false,
+			type: "POST",
+			error :WebUI.handleError,
+			success :WebUI.handleResponse
+		});
+		return false;						// jal 20131107 Was false, but inhibits clicking on radiobutton inside a table in Chrome.
+	},
+
+	prepareAjaxCall: function(id, action, fields) {
+		if (!fields)
+			fields = new Object();
+		// Collect all input, then create input.
+		this.getInputFields(fields);
+		fields.webuia = action;
+		fields.webuic = id;
+		fields["$pt"] = DomUIpageTag;
+		fields["$cid"] = DomUICID;
+
+		return {
+			url :DomUI.getPostURL(),
+			dataType :"*",
+			data :fields,
+			cache :false,
+			type: "POST",
+			success :WebUI.handleResponse,
+			error :WebUI.handleError
+		};
+	},
+
+	scall : function(id, action, fields) {
+		var call = WebUI.prepareAjaxCall(id, action, fields);
+		WebUI.cancelPolling();
+		$.ajax(call);
+	},
+
+	jsoncall: function(id, fields) {
+		if (!fields)
+			fields = new Object();
+		fields.webuia = "$pagejson";
+		fields.webuic = id;
+		fields["$pt"] = DomUIpageTag;
+		fields["$cid"] = DomUICID;
+
+		var response = "";
+		$.ajax( {
+			url :DomUI.getPostURL(),
+			dataType :"text/xml",
+			data :fields,
+			cache :false,
+			async: false,
+			type: "POST",
+			success : function(data, state) {
+				response = data;
+			},
+			error :WebUI.handleError
+		});
+//		console.debug("jsoncall-", response);
+//		try {
+		return eval("("+response+")");
+//		} catch(x) {
+//			console.debug("json data error", x);
+//		}
+	},
+
+	/**
+	 * Send a server request to a component, which will be handled by that component's componentHandleWebAction
+	 * method. The json data is sent as a string parameter with the name "json"; the response is handled as a normal
+	 * DomUI page request: the page is updated and any delta is returned.
+	 * @returns void
+	 */
+	sendJsonAction: function(id, action, json) {
+		var fields = new Object();
+		fields.json = JSON.stringify(json);
+		WebUI.scall(id, action, fields);
+	},
+
+	/**
+	 * Call a JSON handler on a component. This is "out of bound": the current browser state of
+	 * the page is /not/ sent, and the response must be a JSON document which will be the return
+	 * value of this function.
+	 *
+	 * @param id
+	 * @param fields
+	 * @returns
+	 */
+	callJsonFunction: function(id, action, fields) {
+		if (!fields)
+			fields = new Object();
+		fields.webuia = "#"+action;
+		fields.webuic = id;
+		fields["$pt"] = DomUIpageTag;
+		fields["$cid"] = DomUICID;
+
+		var response = "";
+		$.ajax( {
+			url :DomUI.getPostURL(),
+			dataType :"text/xml",
+			data :fields,
+			cache :false,
+			async: false,
+			type: "POST",
+			success : function(data, state) {
+				response = data;
+			},
+			error :WebUI.handleError
+		});
+//		console.debug("jsoncall-", response);
+//		try {
+		return eval("("+response+")");
+//		} catch(x) {
+//			console.debug("json data error", x);
+//		}
+	},
+
+	/**
+	 * Send a server request to a component, which will be handled by that component's componentHandleWebAction
+	 * method. The json data is sent as a string parameter with the name "json"; the response is handled as a normal
+	 * DomUI page request: the page is updated and any delta is returned.
+	 * @returns void
+	 */
+	sendJsonAction: function(id, action, json) {
+		var fields = new Object();
+		fields.json = JSON.stringify(json);
+		WebUI.scall(id, action, fields);
+	},
+
+	/**
+	 * Call a JSON handler on a component. This is "out of bound": the current browser state of
+	 * the page is /not/ sent, and the response must be a JSON document which will be the return
+	 * value of this function.
+	 *
+	 * @param id
+	 * @param fields
+	 * @returns
+	 */
+	callJsonFunction: function(id, action, fields) {
+		if (!fields)
+			fields = new Object();
+		fields.webuia = "#"+action;
+		fields.webuic = id;
+		fields["$pt"] = DomUIpageTag;
+		fields["$cid"] = DomUICID;
+
+		var response = "";
+		$.ajax( {
+			url :DomUI.getPostURL(),
+			dataType :"text/xml",
+			data :fields,
+			cache :false,
+			async: false,
+			type: "POST",
+			success : function(data, state) {
+				response = data;
+			},
+			error :WebUI.handleError
+		});
+//		console.debug("jsoncall-", response);
+//		try {
+		return eval("("+response+")");
+//		} catch(x) {
+//			console.debug("json data error", x);
+//		}
+	},
+
+	jsoncall: function(id, fields) {
+		if (!fields)
+			fields = new Object();
+		fields.webuia = "$pagejson";
+		fields.webuic = id;
+		fields["$pt"] = DomUIpageTag;
+		fields["$cid"] = DomUICID;
+
+		var response = "";
+		$.ajax( {
+			url :DomUI.getPostURL(),
+			dataType :"text/xml",
+			data :fields,
+			cache :false,
+			async: false,
+			type: "POST",
+			success : function(data, state) {
+				response = data;
+			},
+			error :WebUI.handleError
+		});
+//		console.debug("jsoncall-", response);
+//		try {
+		return eval("("+response+")");
+//		} catch(x) {
+//			console.debug("json data error", x);
+//		}
+	},
+
+	clickandchange: function(h, id, evt) {
+		//-- Do not call upward handlers too.
+		if(! evt)
+			evt = window.event;
+		if(evt) {
+			evt.cancelBubble = true;
+			if(evt.stopPropagation)
+				evt.stopPropagation();
+		}
+		WebUI.scall(id, 'clickandvchange');
+	},
+
+	valuechanged: function(h, id) {
+		// FIXME 20100315 jal Temporary fix for bug 680: if a DateInput has a value changed listener the onblur does not execute. So handle it here too.... The fix is horrible and needs generalization.
+		var item = document.getElementById(id);
+		if(item && (item.tagName == "input" || item.tagName == "INPUT") && item.className == "ui-di") {
+			//-- DateInput control: manually call the onblur listener.
+			this.dateInputRepairValueIn(item);
+		}
+
+		// Collect all input, then create input.
+		var fields = new Object();
+		this.getInputFields(fields);
+		fields.webuia = "vchange";
+		fields.webuic = id;
+		fields["$pt"] = DomUIpageTag;
+		fields["$cid"] = DomUICID;
+		WebUI.cancelPolling();
+
+		$.ajax( {
+			url :DomUI.getPostURL(),
+			dataType :"*",
+			data :fields,
+			cache :false,
+			type: "POST",
+			success :WebUI.handleResponse,
+			error :WebUI.handleError
+		});
+	},
+
+	/**
+	 * Handle enter key pressed on keyPress for component with onLookupTyping listener. This needs to be executed on keyPress (was part of keyUp handling), otherwise other global return key listener (returnKeyPress handler) would fire.
+	 */
+	onLookupTypingReturnKeyHandler : function(id, event) {
+		var node = document.getElementById(id);
+		if(!node || node.tagName.toLowerCase() != 'input')
+			return;
+
+		if(!event){
+			event = window.event;
+			if (!event)
+				return;
+		}
+
+		var keyCode = WebUI.normalizeKey(event);
+		var isReturn = (keyCode == 13000 || keyCode == 13);
+
+		if (isReturn) {
+			//cancel current scheduledOnLookupTypingTimerID
+			if (WebUI.scheduledOnLookupTypingTimerID){
+				//cancel already scheduled timer event
+				window.clearTimeout(WebUI.scheduledOnLookupTypingTimerID);
+				WebUI.scheduledOnLookupTypingTimerID = null;
+			}
+			//Do not call upward handlers too, we do not want to trigger on value changed by return pressed.
+			event.cancelBubble = true;
+			if(event.stopPropagation)
+				event.stopPropagation();
+
+			//locate keyword input node
+			var selectedIndex = WebUI.getKeywordPopupSelectedRowIndex(node);
+			var trNode = selectedIndex < 0 ? null : $(node.parentNode).children("div.ui-lui-keyword-popup").children("div").children("table").children("tbody").children("tr:nth-child(" + selectedIndex + ")").get(0);
+			if(trNode){
+				//trigger click on row
+				WebUI.setKeywordPopupSelectedRowIndex(node, -1);
+				$(trNode).trigger('click');
+			} else {
+				//trigger lookupTypingDone when return is pressed
+				WebUI.lookupTypingDone(id);
+			}
+		}
+	},
+
+	/**
+	 * Handle for timer delayed actions, used for onLookupTyping event.
+	 */
+	scheduledOnLookupTypingTimerID: null,
+
+	/*
+	 * Executed as onkeyup event on input field that has implemented listener for onLookupTyping event.
+	 * In case of return key call lookupTypingDone ajax that is transformed into onLookupTyping(done=true).
+	 * In case of other key, lookupTyping funcion is called with delay of 500ms. Previuosly scheduled lookupTyping function is canceled.
+	 * This cause that fast typing would not trigger ajax for each key stroke, only when user stops typing for 500ms ajax would be called by lookupTyping function.
+	 */
+	scheduleOnLookupTypingEvent : function(id, event) {
+		var node = document.getElementById(id);
+		if(!node || node.tagName.toLowerCase() != 'input')
+			return;
+
+		if(!event){
+			event = window.event;
+			if (!event)
+				return;
+		}
+		var keyCode = WebUI.normalizeKey(event);
+		var isReturn = (keyCode == 13000 || keyCode == 13);
+		if (isReturn) { //handled by onLookupTypingReturnKeyHandler, just cancel propagation
+			event.cancelBubble = true;
+			if(event.stopPropagation)
+				event.stopPropagation();
+			return;
+		}
+
+		var isLeftArrowKey = (keyCode == 37000 || keyCode == 37);
+		var isRightArrowKey = (keyCode == 39000 || keyCode == 39);
+		if (isLeftArrowKey || isRightArrowKey){
+			//in case of left or right arrow keys do nothing
+			return;
+		}
+		if (WebUI.scheduledOnLookupTypingTimerID){
+			//cancel already scheduled timer event
+			window.clearTimeout(WebUI.scheduledOnLookupTypingTimerID);
+			WebUI.scheduledOnLookupTypingTimerID = null;
+		}
+		var isDownArrowKey = (keyCode == 40000 || keyCode == 40);
+		var isUpArrowKey = (keyCode == 38000 || keyCode == 38);
+		if (isDownArrowKey || isUpArrowKey) {
+			//Do not call upward handlers too, we do not want to trigger on value changed by return pressed.
+			event.cancelBubble = true;
+			if(event.stopPropagation)
+				event.stopPropagation();
+
+			//locate keyword input node
+			var selectedIndex = WebUI.getKeywordPopupSelectedRowIndex(node);
+			if(selectedIndex < 0)
+				selectedIndex = 0;
+			var trNode = $(node.parentNode).children("div.ui-lui-keyword-popup").children("div").children("table").children("tbody").children("tr:nth-child(" + selectedIndex + ")").get(0);
+			if(trNode){
+				trNode.className = "ui-keyword-popup-row";
+			}
+			var trNodes = $(node.parentNode).children("div.ui-lui-keyword-popup").children("div").children("table").children("tbody").children("tr");
+			if (trNodes.length > 0){
+				var divPopup = $(node.parentNode).children("div.ui-lui-keyword-popup").get(0);
+				if (divPopup){
+					$(divPopup).fadeIn(300);
+					//must be set due to IE bug in rendering
+					node.parentNode.style.zIndex = divPopup.style.zIndex;
+				}
+				if (isDownArrowKey){
+					selectedIndex++;
+				}else{
+					selectedIndex--;
+				}
+				if (selectedIndex > trNodes.length){
+					selectedIndex = 0;
+				}
+				if (selectedIndex < 0){
+					selectedIndex = trNodes.length;
+				}
+				trNode = $(node.parentNode).children("div.ui-lui-keyword-popup").children("div").children("table").children("tbody").children("tr:nth-child(" + selectedIndex + ")").get(0);
+				if(trNode){
+					trNode.className = "ui-keyword-popop-rowsel";
+				}
+			}else{
+				selectedIndex = 0;
+			}
+			WebUI.setKeywordPopupSelectedRowIndex(node, selectedIndex);
+		}
+		else
+			WebUI.scheduledOnLookupTypingTimerID = window.setTimeout("WebUI.lookupTyping('" + id + "')", 500);
+	},
+
+	getKeywordPopupSelectedRowIndex: function(keywordInputNode){
+		var selectedIndexInput = $(keywordInputNode.parentNode).children("input:hidden").get(0);
+		if (selectedIndexInput){
+			if (selectedIndexInput.value && selectedIndexInput.value != ""){
+				return parseInt(selectedIndexInput.value);
+			};
+		}
+		return -1;
+	},
+
+	setKeywordPopupSelectedRowIndex: function(keywordInputNode, intValue){
+		var selectedIndexInput = $(keywordInputNode.parentNode).children("input:hidden").get(0);
+		if (!selectedIndexInput){
+			selectedIndexInput = document.createElement("input");
+			selectedIndexInput.setAttribute("type","hidden");
+			$(keywordInputNode.parentNode).append($(selectedIndexInput));
+		}
+		selectedIndexInput.value = intValue;
+	},
+
+	lookupPopupClicked : function(id) {
+		var node = document.getElementById(id);
+		if(!node || node.tagName.toLowerCase() != 'input') {
+			return;
+		}
+
+		var selectedIndex = WebUI.getKeywordPopupSelectedRowIndex(node);
+		if(selectedIndex < 0)
+			selectedIndex = 0;
+		var trNode = $(node.parentNode).children("div.ui-lui-keyword-popup").children("div").children("table").children("tbody").children("tr:nth-child(" + selectedIndex + ")").get(0);
+		if(trNode){
+			WebUI.clicked(trNode, trNode.id, null);
+		}
+	},
+
+	lookupRowMouseOver : function(keywordInputId, rowNodeId) {
+		var keywordInput = document.getElementById(keywordInputId);
+		if(!keywordInput || keywordInput.tagName.toLowerCase() != 'input') {
+			return;
+		}
+
+		var rowNode = document.getElementById(rowNodeId);
+		if(!rowNode || rowNode.tagName.toLowerCase() != 'tr') {
+			return;
+		}
+
+		var oldIndex = WebUI.getKeywordPopupSelectedRowIndex(keywordInput);
+		if(oldIndex < 0)
+			oldIndex = 0;
+
+		var trNodes = $(rowNode.parentNode).children("tr");
+		var newIndex = 0;
+		for(var i = 1; i <= trNodes.length; i++){
+			if (rowNode == trNodes.get(i-1)) {
+				newIndex = i;
+				break;
+			}
+		}
+
+		if (oldIndex != newIndex){
+			var deselectRow = $(rowNode.parentNode).children("tr:nth-child(" + oldIndex + ")").get(0);
+			if (deselectRow){
+				deselectRow.className = "ui-keyword-popop-row";
+			}
+			rowNode.className = "ui-keyword-popop-rowsel";
+			WebUI.setKeywordPopupSelectedRowIndex(keywordInput, newIndex);
+		}
+	},
+
+	//Called only from onBlur of input node that is used for lookup typing.
+	hideLookupTypingPopup: function(id) {
+		var node = document.getElementById(id);
+		if(!node || node.tagName.toLowerCase() != 'input')
+			return;
+		var divPopup = $(node.parentNode).children("div.ui-lui-keyword-popup").get();
+		if (divPopup){
+			$(divPopup).fadeOut(200);
+		}
+		//fix z-index to one saved in input node
+		if ($.browser.msie){
+			//IE kills event stack (click is canceled) when z index is set during onblur event handler... So, we need to postpone it a bit...
+			window.setTimeout(function() { try { node.parentNode.style.zIndex = node.style.zIndex;} catch (e) { /*just ignore */ } }, 200);
+		}else{
+			//Other browsers dont suffer of this problem, and we can set z index instantly
+			node.parentNode.style.zIndex = node.style.zIndex;
+		}
+	},
+
+	showLookupTypingPopupIfStillFocusedAndFixZIndex: function(id) {
+		var node = document.getElementById(id);
+		if(!node || node.tagName.toLowerCase() != 'input')
+			return;
+		var wasInFocus = node == document.activeElement;
+		var qDivPopup = $(node.parentNode).children("div.ui-lui-keyword-popup");
+		if (qDivPopup.length > 0){
+			var divPopup = qDivPopup.get(0);
+			//z-index correction must be set manually from javascript (because some bug in IE7 -> if set from domui renders incorrectly until page is refreshed?)
+			divPopup.style.zIndex = node.style.zIndex + 1;
+			node.parentNode.style.zIndex = divPopup.style.zIndex;
+		}else{
+			//fix z-index to one saved in input node
+			node.parentNode.style.zIndex = node.style.zIndex;
+		}
+
+		if (wasInFocus && divPopup){
+			//show popup in case that input field still has focus
+			$(divPopup).show();
+		}
+
+		var trNods = $(qDivPopup).children("div").children("table").children("tbody").children("tr");
+		if (trNods && trNods.length > 0) {
+			for(var i=0; i < trNods.length; i++) {
+				var trNod = trNods.get(i);
+				//we need this jquery way of attaching events, if we use trNod.setAttribute("onmouseover",...) it does not work in IE7
+				$(trNod).bind("mouseover", {nodeId: id, trId: trNod.id}, function(event) {
+					WebUI.lookupRowMouseOver(event.data.nodeId, event.data.trId);
+				});
+			}
+		}
+		if (divPopup){
+			$(divPopup).bind("click", {nodeId: id}, function(event) {
+				WebUI.lookupPopupClicked(event.data.nodeId);
+			});
+		}
+	},
+
+	/*
+	 * In case of longer waiting for lookupTyping ajax response show waiting animated marker.
+	 * Function is called with delay of 500ms from ajax.beforeSend method for lookupTyping event.
+	 */
+	displayWaiting: function(id) {
+		var node = document.getElementById(id);
+		if (node){
+			for ( var i = 0; i < node.childNodes.length; i++ ){
+				if (node.childNodes[i].className == 'ui-lui-waiting'){
+					node.childNodes[i].style.display = 'inline';
+				}
+			}
+		}
+	},
+
+	/*
+	 * Hiding waiting animated marker that was shown in case of longer waiting for lookupTyping ajax response.
+	 * Function is called from ajax.completed method for lookupTyping event.
+	 */
+	hideWaiting: function(id) {
+		var node = document.getElementById(id);
+		if (node){
+			for ( var i = 0; i < node.childNodes.length; i++ ){
+				if (node.childNodes[i].className == 'ui-lui-waiting'){
+					node.childNodes[i].style.display = 'none';
+				}
+			}
+		}
+	},
+
+	lookupTyping : function(id) {
+		var lookupField = document.getElementById(id);
+		//check for exsistence, since it is delayed action component can be removed when action is executed.
+		if (lookupField){
+			// Collect all input, then create input.
+			var fields = new Object();
+			this.getInputFields(fields);
+			fields.webuia = "lookupTyping";
+			fields.webuic = id;
+			fields["$pt"] = DomUIpageTag;
+			fields["$cid"] = DomUICID;
+			WebUI.cancelPolling();
+			var displayWaitingTimerID = null;
+
+			$.ajax( {
+				url :DomUI.getPostURL(),
+				dataType :"*",
+				data :fields,
+				cache :false,
+				type: "POST",
+				global: false,
+				beforeSend: function(){
+					// Handle the local beforeSend event
+					var parentDiv = lookupField.parentNode;
+					if (parentDiv){
+						displayWaitingTimerID = window.setTimeout("WebUI.displayWaiting('" + parentDiv.id + "')", 500);
+					}
+				},
+				complete: function(){
+					// Handle the local complete event
+					if (displayWaitingTimerID) {
+						//handle waiting marker
+						window.clearTimeout(displayWaitingTimerID);
+						displayWaitingTimerID = null;
+						var parentDiv = lookupField.parentNode;
+						if (parentDiv) {
+							WebUI.hideWaiting(parentDiv.id);
+						}
+					}
+					//handle received lookupTyping component content
+					WebUI.showLookupTypingPopupIfStillFocusedAndFixZIndex(id);
+					WebUI.doCustomUpdates();
+				},
+
+				success :WebUI.handleResponse,
+				error :WebUI.handleError
+			});
+		}
+	},
+	lookupTypingDone : function(id) {
+		// Collect all input, then create input.
+		var fields = new Object();
+		this.getInputFields(fields);
+		fields.webuia = "lookupTypingDone";
+		fields.webuic = id;
+		fields["$pt"] = DomUIpageTag;
+		fields["$cid"] = DomUICID;
+		WebUI.cancelPolling();
+
+		$.ajax( {
+			url :DomUI.getPostURL(),
+			dataType :"*",
+			data :fields,
+			cache :false,
+			type: "POST",
+			success :WebUI.handleResponse,
+			error :WebUI.handleError
+		});
+	},
+	handleResponse : function(data, state) {
+		WebUI.clearErrorAsy();
+		if (false && window.console && window.console.debug)
+			console.debug("data is ", data);
+		$.webui(data);
+	},
+	handleError : function(request, status, exc) {
+		var txt = request.responseText;
+		if (document.body)
+			document.body.style.cursor = 'default';
+		// alert('Server error: '+status+", len="+txt.length+", val="+txt);
+		if (txt.length == 0) {
+			//-- Firefox fix: if the page is unloading but a request is pending this may cause an status=ERROR for that page. Prevent us from then overwriting the new document....
+			if(status == "error")
+				return;
+
+			txt = "De server is niet bereikbaar 1, status="+status+", "+request.statusText;
+		}
+		document.write(txt);
+		document.close();
+		window.setTimeout('document.body.style.cursor="default"', 1000);
+		return true;
+	},
+
+	_asyalerted: false,
+	_asyDialog: null,
+
+	handleErrorAsy : function(request, status, exc) {
+		if(WebUI._asyalerted) {
+			//-- We're still in error.. Silently redo the poll.
+			WebUI.startPolling(WebUI._pollInterval);
+			return;
+		}
+//		$.dbg("Got into error state - start "+request.responseText);
+		if(status === "abort")
+			return;
+
+		WebUI._asyalerted = true;
+		var txt = request.responseText || "No response - status="+status;
+		if(txt.length > 512)
+			txt = txt.substring(0, 512)+"...";
+		if(txt.length == 0)
+			txt = WebUI._T.sysPollFailMsg+status;
+
+		/*
+		 * As usual there is a problem with error reporting: if the request is aborted because the browser reloads the page
+		 * any pending request is cancelled and comes in here- but with the wrong error code of course. So to prevent us from
+		 * showing an error message: set a timer to show that message 250 milli's later, and hope the stupid browser disables
+		 * that timer.
+		 */
+		setTimeout(function() {
+			if(WebUI._ignoreErrors)
+				return;
+
+			//-- Show an alert error on top of the screen
+			document.body.style.cursor = 'default';
+			var hdr = document.createElement('div');
+			document.body.appendChild(hdr);
+			hdr.className = 'ui-io-blk2';
+			WebUI._asyHider = hdr;
+
+			var ald = document.createElement('div');
+			document.body.appendChild(ald);
+			ald.className = 'ui-ioe-asy';
+			WebUI._asyDialog = ald;
+
+			var d = document.createElement('div');			// Title bar
+			ald.appendChild(d);
+			d.className = "ui-ioe-ttl";
+			d.appendChild(document.createTextNode(WebUI._T.sysPollFailTitle));	// Server unreachable
+
+			d = document.createElement('div');				// Message content
+			ald.appendChild(d);
+			d.className = "ui-ioe-msg";
+			d.appendChild(document.createTextNode(txt));	// Server unreachable
+
+			d = document.createElement('div');				// Message content
+			ald.appendChild(d);
+			d.className = "ui-ioe-msg2";
+
+			var img = document.createElement('div');
+			d.appendChild(img);
+			img.className = "ui-ioe-img";
+			d.appendChild(document.createTextNode(WebUI._T.sysPollFailCont));	// Waiting for the server to return.
+			WebUI.startPolling(WebUI._pollInterval);
+		}, 250);
+	},
+
+	clearErrorAsy: function() {
+//		$.dbg("clear asy called");
+		if(WebUI._asyDialog) {
+			$(WebUI._asyDialog).remove();
+		}
+		if(WebUI._asyHider) {
+			$(WebUI._asyHider).remove();
+		}
+		WebUI._asyDialog = null;
+		WebUI._asyHider = null;
+		WebUI._asyalerted = false;
+	},
+
+	returnKeyPress : function(evt, node) {
+		var keyCode = WebUI.normalizeKey(evt);
+		if (keyCode != 13000 && keyCode != 13)
+			return true;
+		// Be sure 'onblur' is executed before next action (like submit)
+		try {
+			evt.target.onblur(evt);
+		} catch (err) {
+			// Ignore any error in case that target does not support onblur...
+		}
+		WebUI.scall(evt.currentTarget ? evt.currentTarget.id : node.id, 'returnpressed');
+		return false;
+	},
+
+	/***** DateInput control code ****/
+	dateInputCheckInput: function(evt) {
+		if(! evt) {
+			evt = window.event;
+			if(! evt) {
+				return;
+			}
+		}
+		var c = evt.target || evt.srcElement;
+		WebUI.dateInputRepairValueIn(c);
+	},
+
+	dateInputRepairValueIn : function(c) {
+		if(!c)
+			return;
+		var val = c.value;
+
+		if(!val || val.length == 0) // Nothing to see here, please move on.
+			return;
+		Calendar.__init();
+
+		// -- 20130425 jal if this is a date+time thing the value will hold space-separated time, so make sure to split it;
+		var pos = val.indexOf(' ');
+		var timeval = null;
+		if(pos != -1) {
+			// -- Split into trimmed time part and val = only date
+			timeval = $.trim(val.substring(pos + 1));
+			val = $.trim(val.substring(0, pos));
+		}
+
+		// -- Try to decode then reformat the date input
+		var fmt = Calendar._TT["DEF_DATE_FORMAT"];
+		try {
+			var separatorsCount = WebUI.countSeparators(val);
+			if(separatorsCount < 2) {
+				val = WebUI.insertDateSeparators(val, fmt, separatorsCount);
+				var res = Date.parseDate(val, fmt);
+				c.value = res.print(fmt);
+			} else {
+				try{
+					var resultOfConversion = WebUI.parsingOfFormat(val, fmt);
+					var res = Date.parseDate(resultOfConversion, fmt);
+					c.value = res.print(fmt);
+				}catch(x){
+					Date.parseDate(val, fmt);
+				}
+			}
+		} catch(x) {
+			alert(Calendar._TT["INVALID"]);
+		}
+	},
+
+	/**
+	 * Function that checks is format valid after check that input has separators.
+	 */
+
+	parsingOfFormat: function(inputValue, format){
+		// splits to array of alphanumeric "words" from an input (separators are non-alphanumeric characters)
+		var inputValueSplitted = inputValue.match(/(\w+)/g);
+		var formatWithoutPercentCharSplitted = format.replace(/%/g, "").match(/(\w+)/g);
+		var result = "";
+		for(var i = 0; i < formatWithoutPercentCharSplitted.length; i++){
+			switch(formatWithoutPercentCharSplitted[i]){
+				case "d":
+					result = WebUI.formingResultForDayOrMonth(inputValueSplitted[i], result);
+					break;
+				case "m":
+					result = WebUI.formingResultForDayOrMonth(inputValueSplitted[i], result);
+					break;
+				case "Y":
+					result = WebUI.formingResultForYear(inputValueSplitted[i], result);
+					break;
+			}
+		}
+		result = WebUI.insertDateSeparators(result, format);
+		return result;
+	},
+
+	formingResultForDayOrMonth: function(inputValue, result){
+		if(!WebUI.hasFieldInvalidFormat(inputValue)){
+			return result = WebUI.setDayOrMonthFormat(inputValue, result);
+		}
+		else{
+			throw "Invalid date";
+		}
+	},
+
+	formingResultForYear: function(inputValue, result){
+		var VALID_LENGTH_YEAR = 2;
+		if(inputValue.length == VALID_LENGTH_YEAR){
+			return result = WebUI.setYearFormat(inputValue, result);
+		}
+		else{
+			throw "Invalid date";
+		}
+	},
+
+	/**
+	 * Function that checks is format valid of fields day and month.
+	 */
+	hasFieldInvalidFormat: function(inputValue){
+		var MAX_LENGTH = 2;
+		var FORBIDDEN_CHARACTER = "0";
+
+		return (inputValue.length === MAX_LENGTH && (inputValue.charAt(0) === FORBIDDEN_CHARACTER)) || (inputValue.length > MAX_LENGTH);
+	},
+
+	/**
+	 * Function that converts day and month parts of input string that represents date.
+	 */
+	setDayOrMonthFormat: function(inputValue, result){
+		var NEEDED_CHARACTER_DAY_MONTH = "0";
+
+		if(inputValue.length == 1){
+			result += NEEDED_CHARACTER_DAY_MONTH + inputValue;
+		}else {
+			result += inputValue;
+		}
+		return result;
+	},
+
+	/**
+	 * Function that converts year part of input string that represents date.
+	 */
+	setYearFormat: function(inputValue, result){
+		var NEEDED_CHARACTER_YEAR = "20";
+
+		return result += NEEDED_CHARACTER_YEAR + inputValue;
+	},
+
+	/**
+	 *
+	 */
+	showCalendar : function(id, withtime) {
+		var inp = document.getElementById(id);
+		var params = {
+			inputField :inp,
+			eventName :'click',
+			ifFormat :Calendar._TT[withtime ? "DEF_DATETIME_FORMAT"
+				: "DEF_DATE_FORMAT"],
+			daFormat :Calendar._TT["TT_DATE_FORMAT"],
+			singleClick :true,
+			align :'Br',
+			range : [ 1900, 2999 ],
+			weekNumbers :true,
+			showsTime :withtime,
+			timeFormat :"24",
+			electric :false, // jal 20110125 Fixes bug 885- do not update the field when moving to prevent firing the change handler.
+			step :2,
+			position :null,
+			cache :false
+		};
+
+		// -- Try to show the selected date from the input field.
+		var dateFmt = params.inputField ? params.ifFormat : params.daFormat;
+		params.date = Date.parseDate(inp.value, dateFmt);
+
+		var cal = new Calendar(1, params.date, WebUI.onDateSelect, function(
+			cal) {
+			cal.hide();
+			cal.destroy();
+		});
+		cal.showsOtherMonths = false;
+		cal.showsTime = withtime;
+		cal.time24 = true;
+		cal.params = params;
+		// cal.params = params;
+		cal.weekNumbers = true;
+		cal.setRange(params.range[0], params.range[1]);
+		// cal.setDateStatusHandler(params.dateStatusFunc);
+		// cal.getDateText = params.dateText;
+		// if (params.ifFormat) {
+		// cal.setDateFormat(params.ifFormat);
+		// }
+		// if (params.inputField && typeof params.inputField.value == "string")
+		// {
+		// cal.parseDate(params.inputField.value);
+		// }
+		cal.create();
+		cal.refresh();
+		if (!params.position)
+			cal.showAtElement(params.inputField, params.align);
+		else
+			cal.showAt(params.position[0], params.position[1]);
+	},
+
+	onDateSelect : function(cal) {
+		var p = cal.params;
+		var update = (cal.dateClicked || p.electric);
+		if (update && p.inputField) {
+			p.inputField.value = cal.date.print(p.ifFormat);
+			if (typeof p.inputField.onchange == "function" && cal.dateClicked)
+				p.inputField.onchange();
+		}
+		if (update && p.displayArea)
+			p.displayArea.innerHTML = cal.date.print(p.daFormat);
+		if (update && typeof p.onUpdate == "function")
+			p.onUpdate(cal);
+		if (update && p.flat) {
+			if (typeof p.flatCallback == "function")
+				p.flatCallback(cal);
+		}
+		if (update && p.singleClick && cal.dateClicked)
+			cal.callCloseHandler();
+	},
+
+	wtMouseDown : function(e) {
+		alert(e);
+	},
+
+	oddCharAndClickCallback : function(nodeId, clickId) {
+		WebUI.oddChar(document.getElementById(nodeId));
+		document.getElementById(clickId).click();
+	},
+	oddChar : function(obj) {
+		WebUI.toClip(obj.innerHTML);
+	},
+	handleCalendarChanges : function() {
+		// -- find all Calendars
+		var cals = $("div.ui-wa").get();
+		for ( var i = cals.length; --i >= 0;)
+			WebUI.handleCalendar(cals[i]);
+	},
+
+	/** ******************* Scheduler component ************************ */
+	/**
+	 * This handles item placement for a single calendar. It "discovers" the
+	 * calendar divs using fixed css class names, and uses "extra" attributes
+	 * generated on the root divs for the parts to get the calendar data. This
+	 * handles overlapping appointments by dividing a full day lane into
+	 * multiple "ways". A day lane starts with one "way". For every appointment
+	 * added to a day we check if this appointment overlaps another appointment
+	 * on this way[0]. If it does we create a next way (or use it if it already
+	 * exists) and check if the appointment clashes there. We do this until we
+	 * find a way that can accept the appointment without overlaps. We do this
+	 * for all appointments. At the end we have a list of ways for every day. We
+	 * now start rendering the appointments by going through them per way, then
+	 * per appointment. For every appointment on a way we check if there is an
+	 * appointment on a LATER way that overlaps. If not the width of the
+	 * appointment includes all LATER ways that do not overlap.
+	 *
+	 * For this code to work no item may "overlap" a day. The server code takes
+	 * care of splitting long appointments into multiple "items" here.
+	 */
+	handleCalendar : function(caldiv) {
+		// -- TEST calendar object clicky.
+		var cal = new WebUI.Agenda(caldiv);
+		// $(caldiv).mousedown(function(e) {
+		// cal.timeDown(e);
+		// });
+		cal.loadLayout();
+		cal.reposition();
+	},
+
+	getAbsolutePosition : function(obj) {
+		var top = 0, left = 0;
+		while (obj) {
+			top += obj.offsetTop;
+			left += obj.offsetLeft;
+			obj = obj.offsetParent;
+		}
+		return {
+			x :left,
+			y :top
+		};
+	},
+
+	/**
+	 * None of the "standard" JS libraries like Rico or Prototype have code that
+	 * actually <i>works</i> to get the actual <i>page or absolute</i>
+	 * position of elements when scrolling is used. All of them unconditionally
+	 * add scroll offsets to the relative positions but scrolling *will* cause
+	 * items to become *invisible* because they are scrolled out of view. The
+	 * calls here obtain a location for elements taking scrolling into account,
+	 * and they will return null if the item is not visible at all.
+	 */
+	getAbsScrolledPosition : function(el) {
+		// -- Calculate the element's current offseted locations
+		var bx = el.offsetLeft || 0;
+		var by = el.offsetTop || 0;
+		var ex = bx + el.offsetWidth;
+		var ey = by + el.offsetHeight;
+
+		var el = el.parentNode;
+		while (el != null) {
+			if (el.clientHeight != null) {
+				// -- Check the current location within the parent's bounds.
+				if (by < el.scrollTop)
+					by = el.scrollTop;
+				if (bx < el.scrollLeft)
+					bx = el.scrollLeft;
+				if (bx >= ex || by >= ey) // Not visible
+					return null;
+
+				// -- Check the end coordinates.
+				var vey = el.scrollTop + el.clientHeight;
+				var vex = el.scrollLeft + el.clientWidth;
+				if (ex > vex)
+					ex = vex;
+				if (ey > vey)
+					ey = vey;
+				if (by >= ey || bx >= ex) // Past the viewport's bounds?
+					return null;
+
+				// -- This much of the rectangle fits the viewport. Now make the
+				// position absolute within the viewport.
+				by -= el.scrollTop;
+				ey -= el.scrollTop;
+				bx -= el.scrollLeft;
+				ex -= el.scrollLeft;
+
+				by += el.offsetTop;
+				ey += el.offsetTop;
+				bx += el.offsetLeft;
+				ex += el.offsetLeft;
+			}
+			el = el.parentNode;
+		}
+		return {
+			bx :bx,
+			by :by,
+			ex :ex,
+			ey :ey
+		};
+	},
+
+	/** *************** Polling code ************* */
+	startPolling : function(interval) {
+		if(interval < 100 || interval == undefined || interval == null) {
+			alert("Bad interval: "+interval);
+			return;
+		}
+		WebUI._pollInterval = interval;
+		if (WebUI._pollActive)
+			return;
+		WebUI._pollActive = true;
+		WebUI._pollTimer = setTimeout("WebUI.poll()", WebUI._pollInterval);
+	},
+	cancelPolling : function() {
+		if (!WebUI._pollActive)
+			return;
+		clearTimeout(WebUI._pollTimer);
+		WebUI._pollActive = false;
+	},
+
+	poll : function() {
+		WebUI.cancelPolling();
+
+		/*
+		 * Issue a pollasy request using ajax, then handle the result.
+		 */
+		var fields = new Object();
+		fields.webuia = "pollasy";
+		fields["$pt"] = DomUIpageTag;
+		fields["$cid"] = DomUICID;
+
+		$.ajax( {
+			url :window.location.href,
+			dataType :"*", // "text/xml",
+			data :fields,
+			cache :false,
+			global: false, // jal 20091015 prevent block/unblock on polling call.
+			success :WebUI.handleResponse,
+			error :WebUI.handleErrorAsy
+		});
+	},
+
+	/**
+	 * Send Ajax request to the server every 2 minutes. This keeps the session
+	 * alive. The response can contain commands to execute which will indicate
+	 * important events or changes have taken place.
+	 */
+	pingServer: function(timeout) {
+		var url = DomUIappURL + "to.etc.domui.parts.PollInfo.part";
+		var fields= new Object();
+		fields["$pt"] = DomUIpageTag;
+		fields["$cid"] = DomUICID;
+		$.ajax( {
+			url: url,
+			dataType: "*",
+			data: fields,
+			cache: false,
+			global: false, // jal 20091015 prevent block/unblock on polling call.
+			success: function(data, state) {
+				WebUI.executePollCommands(data);
+			},
+			error : function() {
+				//-- Ignore all errors.
+			}
+		});
+		WebUI.startPingServer(timeout);
+	},
+
+	startPingServer: function(timeout) {
+		if(timeout < 60*1000)
+			timeout = 60*1000;
+		setTimeout("WebUI.pingServer("+timeout+")", timeout);
+	},
+
+	executePollCommands: function(data) {
+		// TBD
+	},
+
+	/** Dynamically loading stylesheets and javascript files (Header Contributer delta's) **/
+	/**
+	 * Load the specified stylesheet by creating a script tag and inserting it @ head.
+	 */
+	loadStylesheet: function(path) {
+		var head = document.getElementsByTagName("head")[0];
+		if(! head)
+			throw "Headless document!?";
+		var link = document.createElement('link');
+		link.type = 'text/css';
+		link.rel = 'stylesheet';
+		link.href = path;
+		link.media = 'screen';
+		head.appendChild(link);
+	},
+
+	loadJavascript: function(path) {
+		var head = document.getElementsByTagName("head")[0];
+		if(! head)
+			throw "Headless document!?";
+		var scp = document.createElement('script');
+		scp.type = 'text/javascript';
+		scp.src = path;
+		head.appendChild(scp);
+	},
+
+	/** Prevents default action to be executed if IE11 is detected */
+	preventIE11DefaultAction : function(e){
+		if((navigator.userAgent.match(/Trident\/7\./))){
+			e.preventDefault();
+		}
+	},
+
+	/** ***************** File upload stuff. **************** */
+	fileUploadChange : function(e) {
+		var tgt = e.currentTarget || e.srcElement;
+		var vv = tgt.value.toString();
+
+		if (!vv) { //IE when setting the value to empty..
+			return;
+		}
+
+		// -- Check extensions,
+		var allowed = tgt.getAttribute('fuallowed');
+		if (allowed) {
+			var ok = false;
+			var spl = allowed.split(',');
+
+			vv = vv.toLowerCase();
+			for(var i = 0; i < spl.length; i++) {
+				var ext = spl[i].toLowerCase();
+				if(ext.substring(0, 1) != ".")
+					ext = "."+ext;
+				if(ext == ".*") {
+					ok = true;
+					break;
+				} else if(vv.indexOf(ext, vv.length - ext.length) !== -1) {
+					ok = true;
+					break;
+				}
+			}
+
+			if (!ok) {
+				var parts = vv.split('.');
+				alert(WebUI.format(WebUI._T.uploadType, (parts.length > 1) ? parts.pop() : '', allowed));
+				tgt.value = "";
+				return;
+			}
+		}
+
+		if(typeof tgt.files[0] !== "undefined") {
+			var s = tgt.getAttribute('fumaxsize');
+			var maxSize = 0;
+			try {
+				maxSize = Number(s);
+			} catch(x) {
+			}
+			if(maxSize <= 0 || maxSize === NaN)
+				maxSize = 100 * 1024 * 1024;						// Default max size is 100MB
+
+			var size = tgt.files[0].size;
+			if(size > maxSize) {
+				alert(WebUI.format(WebUI._T.buplTooBig, maxSize));
+				tgt.value = "";
+				return;
+			}
+		}
+
+		// -- Step 2: create or locate an iframe to handle the upload;
+		var iframe = document.getElementById('webuiif');
+		if (iframe) {
+			iframe.parentNode.removeChild(iframe);
+			iframe = undefined;
+		}
+		if (!iframe) {
+			if(jQuery.browser.msie && ! WebUI.isNormalIE9plus()) {			// MicroMorons report ie9 for ie7 emulation of course
+				// -- IE's below 8 of course have trouble. What else.
+				iframe = document
+					.createElement('<iframe name="webuiif" id="webuiif" src="#" style="display:none; width:0; height:0; border:none" onload="WebUI.ieUpdateUpload(event)">');
+				document.body.appendChild(iframe);
+			} else {
+				iframe = document.createElement('iframe');
+				iframe.id = 'webuiif';
+				iframe.name = "webuiif";
+				iframe.src = "#";
+				iframe.style.display = "none";
+				iframe.style.width = "0px";
+				iframe.style.height = "0px";
+				iframe.style.border = "none";
+				iframe.onload = function() {
+					WebUI.updateUpload(iframe.contentDocument);
+				};
+				document.body.appendChild(iframe);
+			}
+		}
+
+		// -- replace the input with a busy indicator
+		var form = tgt.parentNode; // Get form to submit to later on,
+
+		var img = document.createElement('img');
+		img.border = "0";
+		img.src = DomUIProgressURL;
+		form.parentNode.insertBefore(img, form);
+		form.style.display = 'none';
+
+		// -- Target the iframe
+		form.target = "webuiif"; // Fake a new thingy,
+		form.submit(); // Force submit of the thingerydoo
+		WebUI.blockUI();									// since 20160226: block UI during upload
+	},
+
+	/**
+	 * Called for ie upload garbage only, this tries to decode the utter devastating mess that
+	 * ie makes from xml uploads into an iframe in ie8+. Sigh. The main problem with IE is that
+	 * the idiots that built it mess up XML documents sent to it: when the iframe we use receives
+	 * an xml document the idiot translates it into an html document- with formatted tags and minus
+	 * signs before it to collapse them. Instead of just /rendering/ that they actually /generate/
+	 * that as the content document of the iframe- so getting that means you get no XML at all.
+	 * This abomination was more or less fixed in ie9 - but of course they fucked up compatibility mode
+	 * badly.
+	 *
+	 * Everything below ie9
+	 * ====================
+	 * Everything below ie9 has a special property "XMLDocument" attached to the "document" property of
+	 * the window. This property contains the original XML that was used as an XML tree. So for these
+	 * browsers we just get that and move on.
+	 *
+	 * IE9 in native mode
+	 * ==================
+	 * IE9 in native mode does not mess up the iframe content document, so this mode takes the path of
+	 * all browsers worth the name. Since the original problem was solved the XMLDocument property no
+	 * longer exists.
+	 *
+	 * IE9 in compatibility mode (IE7)
+	 * ===============================
+	 * This gets funny. In this mode the browser /still/ messes up the iframe's content model like the
+	 * old IE's it emulates. But of course the XMLDocument property is helpfully removed - EVEN IN THIS
+	 * MODE.
+	 * Remember: this mode is also entered if you are part of a frameset or iframe that is part of an
+	 * mode: the topmost frameset/page determines the mode for the entire set of pages - they are that
+	 * stupid.
+	 *
+	 * I know no other "workaround" than the horrible concoction below; please turn away if you have
+	 * a weak stomach....
+	 *
+	 * In this case we get the "innerText" content of the iframe. This differers from innerHTML in that
+	 * all html tags that IE added are removed, and only the text is retained. Since the html was generated
+	 * by IE in such a way that the xml was presented to the user - this is actually most of our XML...
+	 * But it contains some problems:
+	 * - there is extra whitespace around it's edges, so we need to remove that..
+	 * - All tags start on a new line with a '-' sign before them... Remove all that...
+	 * - The result will have the &amp; entity replaced by & because they are really stupid, again. So replace
+	 *   that too.
+	 * The resulting thing can sometimes be parsed as XML and then processing continues. But it is far
+	 * from perfect. The biggest problem is that the resulting xml has not properly reserved the original
+	 * whitespace; this may lead to rendering problems.
+	 *
+	 * But as far as I know no other solution to this stupifying bug is possible. Please prove me wrong.
+	 *
+	 * @param e
+	 */
+	ieUpdateUpload : function(e) { // Piece of crap
+		var iframe = document.getElementById('webuiif');
+		var xml;
+		if(iframe.contentWindow && iframe.contentWindow.document.XMLDocument) {
+			xml = iframe.contentWindow.document.XMLDocument; // IMPORTANT Fucking MS Crap!!!! See http://p2p.wrox.com/topic.asp?whichpage=1&TOPIC_ID=62981&#153594
+		} else if(iframe.contentDocument) {
+			var crap = iframe.contentDocument.body.innerText;
+			crap = crap.replace(/^\s+|\s+$/g, ''); // trim
+			crap = crap.replace(/(\n|\r)-*/g, ''); // remove '\r\n-'. The dash is optional.
+			crap = crap.replace(/&/g, '&amp;');		// Replace & with entity
+//			alert('crap='+crap);
+			if(window.DOMParser) {
+				var parser = new DOMParser();
+				xml = parser.parseFromString(crap);
+			} else if(window.ActiveXObject) {
+				xml = new ActiveXObject("Microsoft.XMLDOM");
+				if(! xml.loadXML(crap)) {
+					alert('ie9 in emulation mode unfixable bug: cannot parse xml');
+					window.location.href = window.location.href;
+					WebUI.unblockUI();
+					return;
+				}
+			} else {
+				alert('No idea how to parse xml today.');
+			}
+		} else
+			alert('IE error: something again changed in xml source structure of the iframe, sigh');
+
+		WebUI.updateUpload(xml, iframe);
+	},
+	updateUpload : function(doc, ifr) {
+		try {
+			jQuery.fn.executeDeltaXML(doc);
+		} catch (x) {
+			alert(x);
+			throw x;
+		} finally {
+			WebUI.unblockUI();
+		}
+		/*
+		 * 20081015 jal De iframe zou verwijderd moeten worden, maar als ik dat
+		 * doe onder Firefox dan blijft de browser in "busy" mode staan, met een
+		 * rode enabled "stop" knop en een "busy" mouse.. Voor nu hergebruiken
+		 * we de bestaande iframe dan maar en accepteren de verloren resources.
+		 */
+		// iframe.parentNode.removeChild(iframe); // Suicide..
+	},
+
+
+	unloaded : function() {
+		WebUI._ignoreErrors = true;
+		WebUI.sendobituary();
+	},
+
+	beforeUnload: function() {
+		//-- Make sure no "ajax" errors are reported.
+		WebUI._ignoreErrors = true;
+	},
+
+	/**
+	 * Called at page unload time, this quickly tries to send an obituary to the
+	 * server. This is currently unconditional but can later be augmented to
+	 * send the obituary only when the browser window closes.
+	 */
+	sendobituary : function() {
+		try {
+			var rq;
+			if (window.XMLHttpRequest) {
+				rq = new XMLHttpRequest();
+			} else if (window.ActiveXObject) {
+				rq = new ActiveXObject("Microsoft.XMLHTTP");
+			} else {
+				alert("Cannot send obituary (no transport)");
+				return;
+			}
+			rq.open("GET", DomUI.getObituaryURL() + "?$cid=" + DomUICID + "&webuia=OBITUARY&$pt=" + DomUIpageTag, false);
+			rq.send(null);
+		} catch(ex) {
+//			alert("Sending obit failed:"+ex);
+		}
+	},
+
+
+	_selectStart : undefined,
+
+
+	/** ************ Drag-and-drop support code ****************** */
+	/**
+	 * When mouse is downed on an item that is draggable. This moves to PREDRAG
+	 * mode where the first move will create a visible representation of this
+	 * node, ready for dropping.
+	 */
+	dragMouseDown : function(item, evt) {
+		WebUI.dragReset();
+		WebUI._dragType = item.getAttribute('uitype');
+		if (!WebUI._dragType)
+			alert("This DRAGGABLE node has no 'uitype' attribute??");
+		var dragAreaId = item.getAttribute('dragarea');
+		if (dragAreaId){
+			WebUI._dragNode  = document.getElementById(dragAreaId);
+		}else
+			WebUI._dragNode = item;
+		WebUI._dragMode = 1; // PREDRAG
+		$(document.body).bind("mousemove", WebUI.dragMouseMove);
+		$(document.body).bind("mouseup", WebUI.dragMouseUp);
+		var apos = WebUI.getAbsolutePosition(item);
+		WebUI._dragSourceOffset = apos;
+		apos.x = evt.clientX - apos.x;
+		apos.y = evt.clientY - apos.y;
+		if(evt.preventDefault)
+			evt.preventDefault(); // Prevent ffox image dragging
+		else{
+			evt.returnValue = false;
+		}
+		if(document.attachEvent){
+			document.attachEvent( "onselectstart", WebUI.preventSelection);
+		}
+	},
+
+	preventSelection : function(){
+		return false;
+	},
+
+	dragMouseUp : function() {
+		// -- If we're in DRAGGING mode we may accept the drop
+		try {
+			if (WebUI._dragMode == 2) {
+				WebUI.dragClearTimer();
+				var dz = WebUI.dropTargetFind(WebUI._dragLastX,
+					WebUI._dragLastY);
+				if (dz) {
+					WebUI.dropClearZone(); // Discard any dropzone visuals
+					dz._drophandler.drop(dz);
+				}else{
+					WebUI._dragNode.style.display='';//no drop zone, so restore the dragged item
+				}
+			}
+		} finally {
+			WebUI.dragReset();
+		}
+	},
+
+	dragMouseMove : function(e) {
+		if (WebUI._dragMode == 0) {
+			WebUI.dragReset();
+			return;
+		}
+
+		if (WebUI._dragMode == 1) {
+			// -- preDRAG mode: create the node copy, then move it to the
+			// offset' location.
+			WebUI._dragCopy = WebUI.dragCreateCopy(WebUI._dragNode);
+			//MVE make this optional.
+			WebUI._dragNode.style.display='none';
+
+			WebUI._dragMode = 2;
+			document.body.appendChild(WebUI._dragCopy);
+		}
+		WebUI._dragCopy.style.top = (e.clientY - WebUI._dragSourceOffset.y)
+			+ "px";
+		WebUI._dragCopy.style.left = (e.clientX - WebUI._dragSourceOffset.x)
+			+ "px";
+		// console.debug("currentMode: "+WebUI._dragMode+",
+		// type="+WebUI._dragType);
+		WebUI._dragLastX = e.clientX;
+		WebUI._dragLastY = e.clientY;
+		WebUI.dragResetTimer();
+	},
+
+	dragCreateCopy : function(source) {
+		var dv = document.createElement('div');
+
+		// If we drag a TR we need to encapsulate the thingy in a table/tbody to prevent trouble.
+		if(source.tagName != "TR") {
+			dv.innerHTML = source.innerHTML;
+		} else {
+			//-- This IS a tr. Create a table/TBody then add the content model
+			var t = document.createElement('table');
+			dv.appendChild(t);
+			var b = document.createElement('tbody');
+			t.appendChild(b);
+			b.innerHTML = source.innerHTML;			// Copy tr inside tbody we just constructed
+
+			//-- Find parent table's CSS class so we can copy it's style.
+			var dad = WebUI.findParentOfTagName(source, 'TABLE');
+			if(dad) {
+				t.className= dad.className;
+			}
+		}
+
+		dv.style.position = 'absolute';
+		dv.style.width = $(source).width() + "px";
+		dv.style.height = $(source).height() + "px";
+		//console.debug("DragNode isa "+source.tagName+", "+dv.innerHTML);
+		return dv;
+	},
+
+	/**
+	 * Resets the dropzone timer. Called when in DRAGGING mode and the mouse
+	 * moves, this resets any "open" dropzone indicators and resets the timer on
+	 * which drop zone effects are done. This causes the dropzone indicator
+	 * delay when moving the mouse.
+	 */
+	dragResetTimer : function() {
+		WebUI.dragClearTimer();
+		WebUI._dragTimer = setTimeout("WebUI.dragTimerFired()", 250);
+	},
+	dragClearTimer : function() {
+		if (WebUI._dragTimer) {
+			clearTimeout(WebUI._dragTimer);
+			delete WebUI._dragTimer;
+		}
+	},
+
+	/**
+	 * Fires when in DRAGGING mode and the mouse has not moved for a while. It
+	 * initiates the rendering of any drop zone indicators if the mouse is above
+	 * a drop zone.
+	 */
+	dragTimerFired : function() {
+		// console.debug("timer fired");
+		var dz = WebUI.dropTargetFind(WebUI._dragLastX, WebUI._dragLastY);
+		if (!dz) {
+			WebUI.dropClearZone();
+			return;
+		}
+
+		// -- Un-notify the previous dropzone and notify the new'un
+		if (dz == WebUI._currentDropZone) {
+			dz._drophandler.checkRerender(dz);
+			return;
+		}
+		WebUI.dropClearZone();
+		WebUI._currentDropZone = dz;
+		dz._drophandler.hover(dz);
+		// console.debug("AlterClass on "+dz._dropTarget);
+	},
+
+	findDropZoneHandler : function(type) {
+		if (type == "ROW")
+			return WebUI._ROW_DROPZONE_HANDLER;
+		return WebUI._DEFAULT_DROPZONE_HANDLER;
+	},
+
+	dropClearZone : function() {
+		if (WebUI._currentDropZone) {
+			WebUI._currentDropZone._drophandler.unmark(WebUI._currentDropZone);
+			delete WebUI._currentDropZone;
+		}
+	},
+
+	/**
+	 * Clears any node being dragged.
+	 */
+	dragReset : function() {
+		WebUI.dragClearTimer();
+		if (WebUI._dragCopy) {
+			$(WebUI._dragCopy).remove();
+			WebUI._dragCopy = null;
+		}
+		if (WebUI._dragNode) {
+			$(document.body).unbind("mousemove", WebUI.dragMouseMove);
+			$(document.body).unbind("mouseup", WebUI.dragMouseUp);
+			WebUI._dragNode = null;
+		}
+		WebUI.dropClearZone();
+		WebUI._dragMode = 0; // NOTDRAGGED
+
+		if(document.detachEvent){
+			document.detachEvent( "onselectstart", WebUI.preventSelection);
+		}
+
+//		if(WebUI._selectStart){
+//			document.onselectstart = WebUI._selectStart;
+//		}
+	},
+
+	/**
+	 * Gets or recalculates the list of possible drop targets and their absolute
+	 * on-screen position. This list is used to determine if the mouse is "in" a
+	 * drop target. The list gets cached globally in the WebUI object; if an
+	 * AJAX request is done the list gets cleared.
+	 */
+	dropGetList : function() {
+		if (WebUI._dropList)
+			return WebUI._dropList;
+
+		// -- Reconstruct the droplist. Find all objects that possess the
+		// ui-drpbl class.
+		var dl = $(".ui-drpbl").get();
+		WebUI._dropList = new Array();
+		for ( var i = dl.length; --i >= 0;) {
+			var drop = dl[i];
+			var types = drop.getAttribute('uitypes');
+			if (!types)
+				continue;
+			var def = new Object();
+			def._dropTarget = drop; // Store the objects' DOM node,
+			def._position = WebUI.getAbsolutePosition(drop);
+			def._width = drop.clientWidth;
+			def._height = drop.clientHeight;
+			var tar = types.split(",");
+			def._types = tar;
+			def._drophandler = WebUI.findDropZoneHandler(drop
+				.getAttribute('uidropmode'));
+			var id = drop.getAttribute('uidropbody');
+			if (id) {
+				def._tbody = document.getElementById(id);
+				if (!def._tbody) {
+					alert('Internal error: the TBODY ID=' + id + ' cannot be located (row dropTarget)');
+					continue;
+				}
+				WebUI.dropRemoveNonsense(def._tbody);
+			}
+
+			WebUI._dropList.push(def);
+		}
+		return WebUI._dropList;
+	},
+	dropClearList : function() {
+		delete WebUI._dropList;
+	},
+	dropTargetFind : function(x, y) {
+		var dl = WebUI.dropGetList();
+		for ( var i = dl.length; --i >= 0;) {
+			var d = dl[i];
+
+			// -- Contained and of the correct type?
+			if (x >= d._position.x && x < d._position.x + d._width
+				&& y >= d._position.y && y < d._position.y + d._height) {
+				for ( var j = d._types.length; --j >= 0;) {
+					if (d._types[j] == WebUI._dragType)
+						return d;
+				}
+			}
+		}
+		return null;
+	},
+
+	dropRemoveNonsense : function(body) {
+		for ( var i = body.childNodes.length; --i >= 0;) {
+			var n = body.childNodes[i];
+			if (n.nodeName == '#text')
+				body.removeChild(n);
+		}
+	},
+
+
+	checkBrowser: function() {
+		if(this._browserChecked)
+			return;
+		this._browserChecked = true;
+
+		//-- We do not support IE 7 and lower anymore.
+		if($.browser.msie && $.browser.majorVersion < 8) {
+			//-- Did we already report that warning this session?
+			if($.cookie("domuiie") == null) {
+				alert(WebUI.format(WebUI._T.sysUnsupported, $.browser.majorVersion));
+				$.cookie("domuiie", "true", {});
+			}
+		}
+	},
+
+	/** ***************** ScrollableTabPanel stuff. **************** */
+	_ignoreScrollClick: 0,
+
+	scrollLeft : function(bLeft) {
+		if(this._ignoreScrollClick != 0 || $(bLeft).hasClass('ui-stab-dis'))
+			return;
+
+		var scrlNavig = $(bLeft.parentNode);
+		var offset = -1 * parseInt($('ul',scrlNavig).css('marginLeft'));
+		var diff = $(scrlNavig).width() - 2 * $('.ui-stab-scrl-left',scrlNavig).width();
+		var me = this;
+		var disa = false;
+		if ( diff >= offset ){
+			disa = true;
+			diff = offset;
+		}
+		this._ignoreScrollClick++;
+		$('ul',scrlNavig).animate({marginLeft: '+=' + diff}, 400, 'swing', function() {
+			$('.ui-stab-scrl-right', scrlNavig).removeClass('ui-stab-dis');
+			if(disa){
+				$(bLeft).addClass('ui-stab-dis');
+			}
+			me._ignoreScrollClick--;
+		});
+	},
+
+	scrollRight : function(bRight) {
+		if(this._ignoreScrollClick != 0 || $(bRight).hasClass('ui-stab-dis'))
+			return;
+
+		var scrlNavig = $(bRight.parentNode);
+		var tabsTotalWidth = $('li:last',scrlNavig).width() + 8 /* paddong = 8 */ + $('li:last',scrlNavig).offset().left - $('li:first',scrlNavig).offset().left;
+		var tabsVisibleWidth = $(scrlNavig).width() - 2 * $('.ui-stab-scrl-right',scrlNavig).width();
+		var maxLeftOffset = tabsTotalWidth - tabsVisibleWidth;
+		var diff = tabsVisibleWidth;
+		var offset = -1 * parseInt($('ul',scrlNavig).css('marginLeft'));
+
+		var disa = false;
+		if (offset >= maxLeftOffset){
+			return;
+		} else if (diff + offset >= maxLeftOffset){
+			diff = maxLeftOffset - offset;
+			disa = true;
+		}
+		this._ignoreScrollClick++;
+		var me = this;
+		$('ul', scrlNavig ).animate({marginLeft: '-=' + diff},400, 'swing', function() {
+			$('.ui-stab-scrl-left', scrlNavig).removeClass('ui-stab-dis');
+			if (disa){
+				$(bRight).addClass('ui-stab-dis');
+			}
+			me._ignoreScrollClick--;
+		});
+	},
+
+	recalculateScrollers : function(scrlNavigId){
+		var scrlNavig = document.getElementById(scrlNavigId);
+		var tabsTotalWidth = $('li:last',scrlNavig).width() + 8 /* paddong = 8 */ + $('li:last',scrlNavig).offset().left - $('li:first',scrlNavig).offset().left;
+		var tabsVisibleWidth = $(scrlNavig).width() - 2 * $('.ui-stab-scrl-right',scrlNavig).width();
+		//WebUI.debug('debug1', 50, 150, 'width:' + ($('li:last',scrlNavig).width() + 8 + $('li:last',scrlNavig).offset().left - $('li:first',scrlNavig).offset().left));
+		//WebUI.debug('debug2', 50, 200, 'offsetX:' + ($('li:first',scrlNavig).offset().left - $('ul', scrlNavig).offset().left));
+
+		if(tabsTotalWidth > tabsVisibleWidth){
+			var leftM = parseInt($('ul',scrlNavig).css('marginLeft'));
+			//WebUI.debug('debug2', 50, 200, 'leftM:' + leftM);
+			if (tabsTotalWidth + leftM > tabsVisibleWidth){
+				$('.ui-stab-scrl-right',scrlNavig).removeClass('ui-stab-dis');
+			}else{
+				$('.ui-stab-scrl-right',scrlNavig).addClass('ui-stab-dis');
+			}
+			if (leftM < 0){
+				$('.ui-stab-scrl-left',scrlNavig).removeClass('ui-stab-dis');
+			}else{
+				$('.ui-stab-scrl-left',scrlNavig).addClass('ui-stab-dis');
+			}
+		}else{
+
+			// 20121115 btadic: we just want to render scroll buttons without arrows and actions. We don't want empty spaces.
+			//$('.ui-stab-scrl-left',scrlNavig).css('visibility','hidden');
+			//$('.ui-stab-scrl-right',scrlNavig).css('visibility','hidden');
+			$('.ui-stab-scrl-left',scrlNavig).css('display','none');
+			$('.ui-stab-scrl-right',scrlNavig).css('display','none');
+			$('ul', scrlNavig).animate({marginLeft: 0}, 400, 'swing');
+		}
+	},
+
+	/** ***************** Stretch elemnt height. Must be done via javascript. **************** */
+	stretchHeight : function(elemId) {
+		var elem = document.getElementById(elemId);
+		if (!elem){
+			return;
+		}
+		WebUI.stretchHeightOnNode(elem);
+	},
+
+	stretchHeightOnNode : function(elem) {
+		var elemHeight = $(elem).height();
+		var totHeight = 0;
+		$(elem).siblings().each(function(index, node) {
+			//do not count target element and other siblings positioned absolute or relative to parent in order to calculate how much space is actually taken / available
+			if (node != elem && $(node).css('position') == 'static' && ($(node).css('float') == 'none' || $(node).css('width') != '100%' /* count in floaters that occupies total width */)){
+				//In IE7 hidden nodes needs to be additionaly excluded from count...
+				if (!($(node).css('visibility') == 'hidden' || $(node).css('display') == 'none')){
+					//totHeight += node.offsetHeight;
+					totHeight += $(node).outerHeight(true);
+				}
+			}
+		});
+		var elemDeltaHeight = $(elem).outerHeight(true) - $(elem).height(); //we need to also take into account elem paddings, borders... So we take its delta between outter and inner height.
+		if (WebUI.isIE8orIE8c()){
+			//from some reason we need +1 only for IE8!
+			elemDeltaHeight = elemDeltaHeight + 1;
+		}
+		$(elem).height($(elem).parent().height() - totHeight - elemDeltaHeight);
+		if($.browser.msie && $.browser.version.substring(0, 1) == "7"){
+			//we need to special handle another IE7 muddy hack -> extra padding-bottom that is added to table to prevent non-necesarry vertical scrollers
+			if (elem.scrollWidth > elem.offsetWidth){
+				$(elem).height($(elem).height() - 20);
+				//show hidden vertical scroller if it is again needed after height is decreased.
+				if ($(elem).css('overflow-y') == 'hidden'){
+					if (elem.scrollHeight > elem.offsetHeight){
+						$(elem).css({'overflow-y' : 'auto'});
+					}
+				}
+				return;
+			}
+		}
+	},
+
+	/**
+	 * This adds a "resize" listener to the window, and every window "resize" it will call a method
+	 * to recalculate the height of a single div (flexid) based on the position of:
+	 * <ul>
+	 *	<li>The 'bottom' position of the element just above it (get it's top position and add it's height)</li>
+	 *	<li>A bottom 'offset', currently an integer in pixels. It will become a "bottom item" id later, so we can auto-size a div "sandwiched" between two other elements.</li>
+	 * </ul>
+	 * The recalculate code will be called every time the window resizes.
+	 *
+	 * @param topid
+	 * @param flexid
+	 * @param bottom
+	 */
+	autoHeightReset: function(topid, flexid, bottom) {
+		$(window).bind("resize", function() {
+			WebUI.recalculateAutoHeight(topid, flexid, bottom);
+		});
+		WebUI.recalculateAutoHeight(topid, flexid, bottom);
+	},
+
+	recalculateAutoHeight: function(topid, flexid, bottom) {
+		try {
+			var tbot = $(topid).offset().top + $(topid).height();
+			var height = $(window).height() - tbot - bottom;
+			$(flexid).height(height+"px");
+		} catch(x) {
+			//-- Ignore for now
+		}
+	},
+
+	setThreePanelHeight: function(top, middle, bottom) {
+//		try {
+		middle = "#"+middle;
+		var height = $(middle).parent().height();		// Assigned height of the container
+
+		var theight = 0;
+		if(typeof top === "string") {
+			theight = $("#"+top).height();					// Get the end of the "top" element
+		} else if(top) {
+			theight = top;
+		}
+
+		//-- Get height of bottom, if present
+		var bheight = 0;
+		if(typeof bottom === "string") {
+			bheight = $("#"+bottom).height();
+		} else if(bottom) {
+			tbottom = bottom;
+		}
+		height -= theight - bheight;
+		if(height < 0) {
+			height = 0;
+		}
+		$(middle).height(height+"px");
+		$(middle).css({"overflow-y": "auto"});
+//		} catch(x) {
+//			//-- Ignore for now
+//		}
+
+	},
+
+	autoHeight: function(flexid, bottom) {
+		$(window).bind("resize", function() {
+			WebUI.autoHeightRecalc(flexid, bottom);
+		});
+		WebUI.autoHeightRecalc(flexid, bottom);
+	},
+
+	autoHeightRecalc: function(flexid, bottom) {
+		var tbot = $(flexid).offset().top;
+		var height = $(window).height() - tbot - bottom;
+		$(flexid).height(height + "px");
+	},
+
+	notifySizePositionChangedOnId: function(elemId) {
+		var element = document.getElementById(elemId);
+		if (!element){
+			return;
+		}
+		// Send back size information to server
+		var fields = new Object();
+		fields.webuia = "notifyClientPositionAndSize";
+		fields[element.id + "_rect"] = $(element).position().left + "," + $(element).position().top + "," + $(element).width() + "," + $(element).height();
+		fields["window_size"] = window.innerWidth + "," + window.innerHeight;
+		WebUI.scall(element.id, "notifyClientPositionAndSize", fields);
+	},
+
+	notifySizePositionChanged: function(event, ui) {
+		var element = ui.helper.get(0);
+		if (!element){
+			return;
+		}
+		WebUI.notifySizePositionChangedOnId(element.id);
+	},
+
+	/** *************** Debug thingy - it can be used internaly for debuging javascript ;) ************** */
+	debug : function(debugId, posX, posY, debugInfoHtml) {
+		//Be aware that debugId must not start with digit when using FF! Just lost 1 hour to learn this...
+		if ("0123456789".indexOf(debugId.charAt(0)) > -1){
+			alert("debugId(" + debugId+ ") starts with digit! Please use different one!");
+		}
+		var debugPanel = document.getElementById(debugId);
+		if (null == debugPanel){
+			debugPanel = document.createElement(debugId);
+			$(debugPanel).attr('id', debugId);
+			$(debugPanel).css('position', 'absolute');
+			$(debugPanel).css('marginLeft', 0);
+			$(debugPanel).css('marginTop', 0);
+			$(debugPanel).css('background-color', 'yellow');
+			$(debugPanel).css('border', '1px');
+			$(debugPanel).css('z-index', 2000);
+			$(debugPanel).appendTo('body');
+		}
+		$(debugPanel).css('left', posX);
+		$(debugPanel).css('top', posY);
+		$(debugPanel).html(debugInfoHtml);
+	},
+
+	/*-- Printing support --*/
+	_frmIdCounter: 0,
+
+	backgroundPrint : function(url) {
+		try {
+			var frmname = WebUI.priparePrintDiv(url);
+			WebUI.framePrint(frmname);
+		} catch (x) {
+			alert("Failed: " + x);
+		}
+	},
+
+	framePrint : function(frmname){
+		if (jQuery.browser.msie) {
+			WebUI.documentPrintIE(frmname);
+		} else {
+			WebUI.documentPrintNonIE(frmname);
+		}
+	},
+
+	documentPrintIE : function(frmname) {
+		try {
+			var frm = window.frames[frmname];
+			$("#"+frmname).load(function() {
+				try {
+					frm.focus();
+					setTimeout(function() {
+						if (!frm.document.execCommand('print', true, null)){
+							alert('cannot print: '+x);
+						}
+					}, 1000);
+				} catch(x) {
+					alert('cannot print: '+x);
+				}
+			});
+		} catch(x) {
+			alert("Failed: "+x);
+		}
+	},
+
+	documentPrintNonIE : function(frmname) {
+		try {
+			var frm = window.frames[frmname];
+			$("#"+frmname).load(function() {
+				try {
+					frm.focus();
+					setTimeout(function() {
+						frm.print();
+					}, 1000);
+				} catch(x) {
+					alert('cannot print: '+x);
+				}
+			});
+		} catch(x) {
+			alert("Failed: "+x);
+		}
+	},
+
+	priparePrintDiv : function(url){
+		// Create embedded sizeless div to contain the iframe, invisibly.
+		var div = document.getElementById('domuiprif');
+		if (div)
+			div.innerHTML = "";
+		else {
+			div = document.createElement('div');
+			div.id = 'domuiprif';
+			div.className = 'ui-printdiv';
+			document.body.appendChild(div);
+		}
+
+		// -- Create an iframe loading the required thingy.
+		var frmname = "dmuifrm" + (WebUI._frmIdCounter++); // Create unique
+		// name to
+		// circumvent
+		// ffox "print
+		// only once"
+		// bug
+		if (url.trim() !== "") {
+			$(div).html('<iframe id="' + frmname + '" name="' + frmname + '" src="' + url + '">');
+		} else {
+			//well, this is simple element printing, so we have some size limitations
+			$(div).html('<iframe id="' + frmname + '" name="' + frmname + '" width="1000px" height="1000px"/>');
+		}
+		return frmname;
+	},
+
+	/*-- Printing support for simple text messages. Parameter is id of input/textarea tag that contrains text to be printed out. --*/
+	printtext: function (id) {
+		var item = document.getElementById(id);
+		var textData;
+		if(item && (item.tagName == "input" || item.tagName == "INPUT" || item.tagName == "textarea" || item.tagName == "TEXTAREA")) {
+			textData = item.value;
+		}
+		if (textData){
+			try {
+				// Create embedded sizeless div to contain the iframe, invisibly.
+				var div = document.getElementById('domuiprif');
+				if(div)
+					div.innerHTML = "";
+				else {
+					div = document.createElement('div');
+					div.id = 'domuiprif';
+					div.className = 'ui-printdiv';
+					document.body.appendChild(div);
+				}
+
+				//-- Create an iframe loading the required thingy.
+				var frmname = "dmuifrm"+(WebUI._frmIdCounter++);		// Create unique name to circumvent ffox "print only once" bug
+
+				$(div).html('<iframe id="'+frmname+'" name="'+frmname+'" width="1000px" height="1000px"/>'); //well, this is simple text printing, so we have some size limitations ;)
+				var frm = window.frames[frmname];
+				frm.document.open();
+				frm.document.write('<html></head></head><body style="margin:0px;">');
+				frm.document.write(textData.replace(/\n/g, '<br/>'));
+				frm.document.write('</body></html>');
+				frm.document.close();
+				frm.focus();
+				frm.print();
+			} catch(x) {
+				alert("Failed: "+x);
+			}
+		}
+	},
+
+	nearestID: function(elem) {
+		while(elem) {
+			if(elem.id)
+				return elem.id;
+			elem = elem.parentNode;
+		}
+		return undefined;
+	},
+
+	handleDevelopmentMode: function() {
+		$(document).bind("keydown", function(e) {
+			if(e.keyCode != 192)
+				return;
+
+			var t = new Date().getTime();
+			if(! WebUI._debugLastKeypress || (t - WebUI._debugLastKeypress) > 250) {
+				WebUI._debugLastKeypress = t;
+				return;
+			}
+
+			//-- Send a DEBUG command to the server, indicating the current node below the last mouse move....
+			var id = WebUI.nearestID(WebUI._debugMouseTarget);
+			if(! id) {
+				id = document.body.id;
+			}
+
+			WebUI.scall(id, "DEVTREE", {});
+		});
+		$(document.body).bind("mousemove", function(e) {
+//			if(WebUI._NOMOVE)
+//				return;
+//			console.debug("move ", e);
+			WebUI._debugMouseTarget = e.target; // e.srcElement || e.originalTarget;
+		});
+	},
+
+
+	/***** Popup menu code *****/
+
+	_popinCloseList: [],
+
+	popupMenuShow: function(refid, menu) {
+		WebUI.registerPopinClose(menu);
+		var pos = $(refid).offset();
+		var eWidth = $(refid).outerWidth();
+		var mwidth = $(menu).outerWidth();
+		var left = (pos.left);
+		if(left + mwidth > screen.width)
+			left = screen.width - mwidth - 10;
+		var top = 3+pos.top;
+		$(menu).css( {
+			position: 'absolute',
+			zIndex: 100,
+			left: left+"px",
+			top: top+"px"
+		});
+
+		$(menu).hide().fadeIn();
+	},
+
+	popupSubmenuShow: function(parentId, submenu) {
+		$(submenu).position({my: 'left top', at: 'center top', of: parentId});
+	},
+
+	/**
+	 * Register the popup. If the mouse leaves the popup window the popup needs to send a POPINCLOSE? command; this
+	 * will tell DomUI server that the popin needs to go. If an item inside the popin is clicked it should mean the
+	 * popin closes too; at that point we will deregister the mouse listener to prevent sending double events.
+	 *
+	 * @param id
+	 */
+	registerPopinClose: function(id) {
+		WebUI._popinCloseList.push(id);
+		$(id).bind("mouseleave", WebUI.popinMouseClose);
+		if(WebUI._popinCloseList.length != 1)
+			return;
+		$(document.body).bind("keydown", WebUI.popinKeyClose);
+//		$(document.body).bind("beforeclick", WebUI.popinBeforeClick);	// Called when a click is done somewhere - not needed anymore, handled from java
+	},
+
+	popinClosed: function(id) {
+		for(var i = 0; i < WebUI._popinCloseList.length; i++) {
+			if(id === WebUI._popinCloseList[i]) {
+				//-- This one is done -> remove mouse handler.
+				$(id).unbind("mousedown", WebUI.popinMouseClose);
+				WebUI._popinCloseList.splice(i, 1);
+				if(WebUI._popinCloseList.length == 0) {
+					$(document.body).unbind("keydown", WebUI.popinKeyClose);
+					$(document.body).unbind("beforeclick", WebUI.popinBeforeClick);
+				}
+				return;
+			}
+		}
+	},
+
+	popinBeforeClick: function(ee1, obj, clickevt) {
+		for(var i = 0; i < WebUI._popinCloseList.length; i++) {
+			var id = WebUI._popinCloseList[i];
+			obj = $(obj);
+			var cl = obj.closest(id);
+			if(cl.size() > 0) {
+				//-- This one is done -> remove mouse handler.
+				$(id).unbind("mousedown", WebUI.popinMouseClose);
+				WebUI._popinCloseList.splice(i, 1);
+				if(WebUI._popinCloseList.length == 0) {
+					$(document.body).unbind("keydown", WebUI.popinKeyClose);
+					$(document.body).unbind("beforeclick", WebUI.popinBeforeClick);
+				}
+				return;
+			}
+		}
+	},
+
+	popinMouseClose: function() {
+		if(WebUI.isUIBlocked())							// We will get a LEAVE if the UI blocks during menu code... Ignore it
+			return;
+
+		try {
+			for(var i = 0; i < WebUI._popinCloseList.length; i++) {
+				var id = WebUI._popinCloseList[i];
+				var el = $(id);
+				if(el) {
+					el.unbind("mousedown", WebUI.popinMouseClose);
+					WebUI.scall(id.substring(1), "POPINCLOSE?", {});
+				}
+			}
+		} finally {
+			WebUI._popinCloseList = [];
+//			$(document.body).unbind("mousedown", WebUI.popinMouseClose);
+			$(document.body).unbind("keydown", WebUI.popinKeyClose);
+			$(document.body).unbind("beforeclick", WebUI.popinBeforeClick);
+		}
+	},
+	popinKeyClose: function(evt) {
+		if(! evt)
+			evt = window.event;
+		var kk = WebUI.normalizeKey(evt);
+		if(kk == 27 || kk == 27000) {
+			// Prevent ESC from cancelling the AJAX call in Firefox!!
+			evt.preventDefault();
+			evt.cancelBubble = true;
+			if(evt.stopPropagation)
+				evt.stopPropagation();
+			WebUI.popinMouseClose();
+		}
+	},
+
+	//We need to re-show element to force IE7 browser to recalculate correct height of element. This must be done to fix some IE7 missbehaviors.
+	refreshElement: function(id) {
+		var elem = document.getElementById(id);
+		if (elem){
+			$(elem).hide();
+			$(elem).show(1); //needs to be done on timeout/animation, otherwise it still fails to recalculate...
+		}
+	},
+
+	//Use this to make sure that item would be visible inside parent scrollable area. It uses scroll animation. In case when item is already in visible part, we just do single blink to gets user attention ;)
+	scrollMeToTop: function(elemId, selColor, offset) {
+		var elem = document.getElementById(elemId);
+		if (!elem){
+			return;
+		}
+		var parent = elem.parentNode;
+		if (!parent){
+			return;
+		}
+		if (parent.scrollHeight > parent.offsetHeight){ //if parent has scroll
+			var elemPos = $(elem).position().top;
+			if (elemPos > 0 && elemPos < parent.offsetHeight){
+				//if elem already visible -> just do one blink
+				if (selColor){
+					var oldColor = $(elem).css('background-color');
+					$(elem).animate({backgroundColor: selColor}, "slow", function(){$(elem).animate({backgroundColor: oldColor}, "fast");});
+				}
+			}else{
+				//else scroll parent to show me at top
+				var newPos = $(elem).position().top + parent.scrollTop;
+				if($.browser.msie && parseInt($.browser.version) < 11){
+					if ($(elem).height() == 0){
+						newPos = newPos - 15; //On IE browsers older than 11 we need this correction :¬|
+					}
+				}
+				if (offset){
+					newPos = newPos - offset;
+				}
+				$(parent).animate({scrollTop: newPos}, 'slow');
+			}
+		}
+	},
+
+	//Use this to make sure that option in dropdown would be visible. It needs fix only in FF sinve IE would always make visible selected option.
+	makeOptionVisible: function(elemId, offset) {
+		if($.browser.msie){
+			//IE already fix this... we need fix only for FF and other browsers
+			return;
+		}
+		var elem = document.getElementById(elemId);
+		if (!elem){
+			return;
+		}
+		var parent = elem.parentNode;
+		if (!parent){
+			return;
+		}
+		if (parent.scrollHeight > parent.offsetHeight){ //if parent has scroll
+			var elemPos = $(elem).position().top;
+			//if elem is not currenlty visible
+			if (elemPos <= 0 || elemPos >= parent.offsetHeight){
+				//else scroll parent to show me at top
+				var newPos = elemPos + parent.scrollTop;
+				if (offset){
+					newPos = newPos - offset;
+				}
+				$(parent).animate({scrollTop: newPos}, 'slow');
+			}
+		}
+	},
+
+	//Returns T if browser is really using IE7 rendering engine (since IE8 compatibility mode presents  browser as version 7 but renders as IE8!)
+	isReallyIE7: function() {
+		//Stupid IE8 in compatibility mode lies that it is IE7, and renders as IE8! At least we can detect that using document.documentMode (it is 8 in that case)
+		//document.documentMode == 7 		 --- IE8 running in IE7 mode
+		//document.documentMode == 8 		 --- IE8 running in IE8 mode or IE7 Compatibility mode
+		//document.documentMode == undefined --- plain old IE7
+		return ($.browser.msie && parseInt($.browser.version) == 7 && (!document.documentMode || document.documentMode == 7));
+	},
+	//Returns T if browser is IE8 or IE8 compatibility mode
+	isIE8orIE8c: function() {
+		//Stupid IE8 in compatibility mode lies that it is IE7, and renders as IE8! At least we can detect that using document.documentMode (it is 8 in that case)
+		//document.documentMode == 7 		 --- IE8 running in IE7 mode
+		//document.documentMode == 8 		 --- IE8 running in IE8 mode or IE7 Compatibility mode
+		//document.documentMode == undefined --- plain old IE7
+		return ($.browser.msie && (parseInt($.browser.version) == 8 || (parseInt($.browser.version) == 7 && document.documentMode == 8)));
+	},
+	//Returns T if browser is IE of at least version 9 and does not run in any of compatibility modes for earlier versions
+	isNormalIE9plus: function() {
+		return ($.browser.msie && parseInt($.browser.version) >= 9 && document.documentMode >= 9);
+	},
+
+	//Returns T if browser is IE of at least version 8 even if it runs in IE7 compatibility mode
+	isIE8orNewer: function() {
+		return ($.browser.msie && (parseInt($.browser.version) >= 8 || (parseInt($.browser.version) == 7 && document.documentMode >= 8)));
+	},
+
+	// CK editor support, map of key (id of editor) value (pair of [editor instance, assigned resize function])
+	_ckEditorMap : {},
+
+	/**
+	 * Register ckeditor for extra handling that is set in CKeditor_OnComplete.
+	 *
+	 * @param id
+	 * @param ckeInstance
+	 */
+	registerCkEditorId : function(id, ckeInstance) {
+		WebUI._ckEditorMap[id] = [ckeInstance, null];
+	},
+
+	/**
+	 * Unregister ckeditor and removes handlings bound to it.
+	 *
+	 * @param id
+	 */
+	unregisterCkEditorId : function(id) {
+		try {
+			var editorBindings = WebUI._ckEditorMap[id];
+			WebUI._ckEditorMap[id] = null;
+			if (editorBindings && editorBindings[1]){
+				$(window).unbind('resize', editorBindings[1]);
+			}
+		} catch (ex) {
+			WebUI.log('error in unregisterCkEditorId: ' + ex);
+		}
+	},
+
+	/**
+	 * Piece of support needed for CK editor to properly fix its size, using _OnComplete handler.
+	 *
+	 * @param id
+	 */
+	CKeditor_OnComplete : function(id) {
+		WebUI.doCustomUpdates();
+		var elem = document.getElementById(id);
+		var parentDiv = elem.parentNode;
+		var editor = WebUI._ckEditorMap[id][0];
+		var resizeFunction = function(ev) {
+			try{
+				editor.resize($(parentDiv).width() - 2, $(parentDiv).height());
+			}catch (ex){
+				WebUI.log('error in CKeditor_OnComplete#resizeFunction: ' + ex);
+			}
+		};
+		WebUI._ckEditorMap[id] = [editor, resizeFunction];
+		$(window).bind('resize', resizeFunction);
+		$(window).trigger('resize');
+	},
+
+	// connects input to usually hidden list select and provides autocomplete
+	// feature inside input. Down arrow does show and focus select list.
+	initAutocomplete : function(inputId, selectId) {
+		var input = document.getElementById(inputId);
+		var select = document.getElementById(selectId);
+		$(input).keyup(function(event) {
+			WebUI.autocomplete(event, inputId, selectId);
+		});
+		$(select).keypress(function(event) {
+			//esc hides select and prevents fireing of click and blur handlers that are temporary disconnected while focus moves back to input
+			var keyCode = WebUI.normalizeKey(event);
+			if (keyCode == 27 || keyCode == 27000) {
+				var oldVal = input.value;
+				var selectOnClick = select.click;
+				var selectOnBlur = select.blur;
+				select.click = null;
+				select.blur = null;
+				select.style.display = 'none';
+				input.focus();
+				input.value = oldVal;
+				select.click = selectOnClick;
+				select.blur = selectOnBlur;
+			}
+		});
+	},
+
+	// does autocomplete part of logic
+	autocomplete : function (event, inputId, selectId) {
+		var select = document.getElementById(selectId);
+		var cursorKeys = "8;46;37;38;39;40;33;34;35;36;45;";
+		if (cursorKeys.indexOf(event.keyCode + ";") == -1) {
+			var input = document.getElementById(inputId);
+			var found = false;
+			var foundAtIndex = -1;
+			for (var i = 0; i < select.options.length; i++){
+				if ((found = select.options[i].text.toUpperCase().indexOf(input.value.toUpperCase()) == 0)){
+					foundAtIndex = i;
+					break;
+				}
+			}
+			select.selectedIndex = foundAtIndex;
+
+			var oldValue = input.value;
+			var newValue = found ? select.options[foundAtIndex].text : oldValue;
+			if (newValue != oldValue) {
+				if (typeof input.selectionStart != "undefined") {
+					//normal browsers
+					input.value = newValue;
+					input.selectionStart = oldValue.length;
+					input.selectionEnd =  newValue.length;
+					input.focus();
+				}
+				if (document.selection && document.selection.createRange) {
+					//IE9
+					input.value = newValue;
+					input.focus();
+					input.select();
+					var range = document.selection.createRange();
+					range.collapse(true);
+					range.moveStart("character", oldValue.length);
+					range.moveEnd("character", newValue.length);
+					range.select();
+				}else if (input.createTextRange) {
+					//IE8-
+					input.value = newValue;
+					var rNew = input.createTextRange();
+					rNew.moveStart('character', oldValue.length);
+					rNew.select();
+				}
+			}
+		}else if (event.keyCode == 40){
+			select.style.display = 'inline';
+			select.focus();
+		}
+	},
+
+	//alignment methods
+	//sticks top of element with nodeId to bottom of element with alignToId, with extra offsetY.
+	alignTopToBottom : function (nodeId, alignToId, offsetY, doCallback){
+		var alignNode = $('#' + alignToId);
+		var node = $('#' + nodeId);
+		var myTopPos;
+		if (node.css('position') == 'fixed'){
+			myTopPos = $(alignNode).offset().top - $(document).scrollTop() + offsetY + $(alignNode).outerHeight(true);
+		}else{
+			myTopPos = $(alignNode).position().top + offsetY + $(alignNode).outerHeight(true);
+		}
+		$(node).css('top', myTopPos);
+		if (doCallback){
+			WebUI.notifySizePositionChangedOnId(nodeId);
+		}
+	},
+
+	//align top of element with nodeId to top of element with alignToId, with extra offsetY
+	alignToTop : function (nodeId, alignToId, offsetY, doCallback){
+		var alignNode = $('#' + alignToId);
+		var node = $('#' + nodeId);
+		var myTopPos;
+		if (node.css('position') == 'fixed'){
+			myTopPos = $(alignNode).offset().top - $(document).scrollTop() + offsetY;
+		}else{
+			myTopPos = $(alignNode).position().top + offsetY;
+		}
+		var nodeHeight = $(node).outerHeight(true);
+		if (myTopPos + nodeHeight > $(window).height()){
+			myTopPos = $(window).height() - nodeHeight;
+		}
+		$(node).css('top', myTopPos);
+		if (doCallback){
+			WebUI.notifySizePositionChangedOnId(nodeId);
+		}
+	},
+
+	//align left edge of element with nodeId to left edge of element with alignToId, with extra offsetX
+	alignToLeft : function (nodeId, alignToId, offsetX, doCallback){
+		var node = $('#' + nodeId);
+		var alignNode = $('#' + alignToId);
+		var myLeftPos;
+		if (node.css('position') == 'fixed'){
+			myLeftPos = $(alignNode).offset().left - $(document).scrollLeft() + offsetX;
+		}else{
+			myLeftPos = $(alignNode).position().left + offsetX;
+		}
+		var nodeWidth = $(node).outerWidth(true);
+		if (myLeftPos + nodeWidth > $(window).width()){
+			myLeftPos = $(window).width() - nodeWidth;
+			if (myLeftPos < 1){
+				myLeftPos = 1;
+			}
+		}
+		$(node).css('left', myLeftPos);
+		if (doCallback){
+			WebUI.notifySizePositionChangedOnId(nodeId);
+		}
+	},
+
+	//align right edge of element with nodeId to right edge of element with alignToId, with extra offsetX
+	alignToRight : function (nodeId, alignToId, offsetX, doCallback){
+		var node = $('#' + nodeId);
+		var alignNode = $('#' + alignToId);
+		var myLeftPos;
+		if (node.css('position') == 'fixed'){
+			myLeftPos = $(alignNode).offset().left - $(document).scrollLeft() + offsetX - $(node).outerWidth(true) + $(alignNode).outerWidth(true) - 3;
+		}else{
+			myLeftPos = $(alignNode).position().left + offsetX - $(node).outerWidth(true) + $(alignNode).outerWidth(true) - 3;
+		}
+		if (myLeftPos < 1){
+			myLeftPos = 1;
+		}
+		$(node).css('left', myLeftPos);
+		if (doCallback){
+			WebUI.notifySizePositionChangedOnId(nodeId);
+		}
+	},
+
+	//align horizontaly middle of element with nodeId to middle of element with alignToId, with extra offsetX
+	alignToMiddle : function (nodeId, alignToId, offsetX, doCallback){
+		var node = $('#' + nodeId);
+		var alignNode = $('#' + alignToId);
+		var myLeftPos;
+		if (node.css('position') == 'fixed'){
+			myLeftPos = $(alignNode).offset().left - $(document).scrollLeft() + ($(alignNode).outerWidth(true) / 2) - ($(node).outerWidth(true) / 2);
+		}else{
+			myLeftPos = $(alignNode).position().left + ($(alignNode).outerWidth(true) / 2) - ($(node).outerWidth(true) / 2);
+		}
+		if (myLeftPos < 1){
+			myLeftPos = 1;
+		}
+		$(node).css('left', myLeftPos);
+		if (doCallback){
+			WebUI.notifySizePositionChangedOnId(nodeId);
+		}
+	}
+});
+
 WebUI._DEFAULT_DROPZONE_HANDLER = {
 	checkRerender : function(dz) {
 	},
