+++ to.etc.domui/src/main/resources/resources/ts/domui.jquery.d.ts
@@ -1,10 +1,11 @@
 interface JQuery {
 	webui(xml: any) : void;
-
+	disableSelection(): void;
+	enableSelection():  void;
 }
 
 interface JQueryStatic {
 	expr: any;
 	browser: any;
-
+	dbg: any;
 }
